+++ s/features/op/op.ts
@@ -1,26 +1,14 @@
 
 import {pub} from "@e280/stz"
-import {signal} from "@e280/strata/signals"
+import {Signal, signal} from "@e280/strata/signals"
 
-import {pod} from "./pod.js"
+import {podium} from "./podium.js"
 import {Pod, PodSelect} from "./types.js"
 
 export class Op<V> {
-	static loading<V>() {
-		return new this<V>()
-	}
-
-	static ready<V>(value: V) {
-		const op = new this<V>()
-		op.signal(["ready", value])
-		return op
-	}
-
-	static error<V>(error: any) {
-		const op = new this<V>()
-		op.signal(["error", error])
-		return op
-	}
+	static loading<V>() { return new this<V>() }
+	static ready<V>(value: V) { return new this<V>(["ready", value]) }
+	static error<V>(error: any) { return new this<V>(["error", error]) }
 
 	static promise<V>(promise: Promise<V>) {
 		const op = new this<V>()
@@ -32,9 +20,19 @@ export class Op<V> {
 		return this.promise(fn())
 	}
 
-	readonly signal = signal<Pod<V>>(["loading"])
+	static all<V>(...ops: Op<V>[]) {
+		const pods = ops.map(op => op.pod)
+		const pod = podium.all(...pods)
+		return new this(pod)
+	}
+
+	readonly signal: Signal<Pod<V>>
 	#resolve = pub<[V]>()
 	#reject = pub<[any]>()
+
+	constructor(pod: Pod<V> = ["loading"]) {
+		this.signal = signal<Pod<V>>(pod)
+	}
 
 	get wait() {
 		return new Promise((resolve, reject) => {
@@ -77,16 +75,20 @@ export class Op<V> {
 		return this.signal()
 	}
 
+	set pod(p: Pod<V>) {
+		this.signal(p)
+	}
+
 	get status() {
 		return this.signal()[0]
 	}
 
 	get value() {
-		return pod.value(this.signal())
+		return podium.value(this.signal())
 	}
 
 	get error() {
-		return pod.error(this.signal())
+		return podium.error(this.signal())
 	}
 
 	get isLoading() {
@@ -108,7 +110,11 @@ export class Op<V> {
 	}
 
 	select<R>(select: PodSelect<V, R>) {
-		return pod.select(this.signal(), select)
+		return podium.select(this.signal(), select)
+	}
+
+	morph<V2>(fn: (value: V) => V2) {
+		return new Op(podium.morph(this.pod, fn))
 	}
 }